JS: DOM API 9
Forms are an important part of working with the DOM in JavaScript. They allow the user to enter data and send it to the server for processing. Interaction with forms includes creating, modifying, receiving and submitting data. Here are some basic operations related to working with forms using JavaScript and the DOM:

Getting a link to a form:
use querySelector to find the first form element on the page.
Getting the value of an input field: form.elements.inputName.value
Setting the input field value: form.elements.inputName.value = 'New Value';
Getting the value of the selected <select> element: form.elements.selectName.value;
Setting the selected value of the <select> element: form.elements.selectName.value = 'optionValue';
Getting the value of the selected elements <input type="checkbox">: form.elements.checkboxName;