JS: DOM API 10
Intercepting and bubbling are two important concepts in the JavaScript DOM API that relate to event handling.
Capturing and bubbling are mechanisms that determine the order in which events propagate in the DOM tree from parent to child (capturing) and from child to parent (bubbling).
When an event occurs on an element, it first goes through the catching phase, starting at the topmost parent element and moving down to the target element. Then, upon reaching the target element, the event goes through a bubbling phase, propagating upward from the target element to the parent elements.