Regular Expressions 5
Quantification in regular expressions allows you to specify the number of repetitions for the previous element or group of elements. It specifies how many times the previous element or group of elements must occur before a match is found.
The following are the most common quantification symbols:
1. * (asterisk): Matches 0 or more repetitions of the previous element. For example, a* will match "a", "aa", "aaa", etc.
2. + (plus): Matches 1 or more repetitions of the previous element. For example, a+ will match "a", "aa", "aaa", etc., but will not match the empty string.
3. ? (question mark): Matches 0 or 1 repetition of the previous element. For example, ab? will match "a" and "ab".
4. {n}: Matches exactly n repetitions of the previous element. For example, a{3} will only match "aaa".
5. {n,}: Matches n or more repetitions of the previous element. For example, a{2,} will match "aa", "aaa", "aaaa", etc.
6. {n,m}: Matches n to m repetitions of the previous element. For example, a{2,4} will match "aa", "aaa", and "aaaa".
Quantifiers can be applied to characters, character classes, groups, and metacharacters. They allow you to more flexibly set patterns and find matches in the text.