Regular Expressions 6
Greed in regular expressions refers to the way a pattern is matched against text. Greedy quantifiers try to match as many characters as possible to satisfy the pattern, even if that means matching more text than necessary. They strive for the maximum expansion of the comparison.
Greedy quantifiers are usually represented by *, +, or ? after the pattern. Here are some examples:
.* - Matches any number of characters (including zero or more), trying to match as many characters as possible.
.+ - Matches one or more characters, trying to match as many characters as possible.
.? - Matches zero or one character, trying to match as many characters as possible.
Greed in regular expressions can be useful when you want to get the widest possible match, but sometimes it can lead to unexpected results. Therefore, it is important to understand how greed works and use it wisely depending on the mapping required.