Regular Expressions 7
In regular expressions, grouping is used to create subexpressions that can be reused or to apply operations on a set of characters. Grouping puts a set of characters inside brackets ( ), and everything inside the brackets is treated as one.
Grouping allows operations to be applied to a set of characters as a single element, and also provides feedback or reference to a group from another part of the regular expression.
Backtracking in regular expressions allows you to refer to the contents of a previously defined group in a regular expression. This is done with a backreference \ followed by the group number. A backreference allows you to reuse a found group in a regular expression so that matches are matched.

1. Grouping without feedback:
Regular expression: (ab)+
Designation: One or more consecutive "ab" groups
Example:
String: "ababab"
Matches: "ab", "ab", "ab"

2. Grouping with feedback:
Regular expression: (\d{2})-\1
Designation: Two digits, then a hyphen, then repeating the contents of the first group
Example:
String: "12-12"
Match: "12-12"
String: "12-34"
No match
Backtracking allows you to refer to the contents of a group defined earlier in a regular expression, making it a powerful tool for more complex matches and substitutions in text.