Regular Expressions 9
In regular expressions, positive lookahead and positive lookbehind allow you to check for certain patterns before or after the current position in the text, but without including those patterns in the actual match result.
Looking ahead (positive lookahead) is done using (?=...). For example, foo(?=bar) will match the substring "foo" only if it is followed by "bar". In this case, "bar" will not be included in the matching result.
Looking back (positive lookbehind) is done using (?<=...). For example, (?<=foo)bar will only match the substring "bar" if it is preceded by "foo". In this case, "foo" will not be included in the match result.
If we want to check that some pattern does not fit, then instead of "=" we put "!".