Regular Expressions 10
Conditional searching in regular expressions is done using special constructs called "positive lookahead" ((?=...)) and "negative lookahead" ((?!...)) to test conditions that must be met. the next part of the line.
Positive lookahead ((?=...)) is used to check that a certain condition matches the characters following the current position. If the condition is met, the search continues. But the result of a positive lookahead match is not included in the final match.
Negative lookahead ((?!...)) is used to test that a certain condition does not match the characters following the current position. If the condition is not met, the search continues. The result of a negative lookahead match is also not included in the final match.
These constructs allow you to specify additional search conditions without including them in the actual match. Thus, they allow you to perform a conditional search without including the matching characters in the result.