JS: Frontend Architecture 1
Application state in the JavaScript front-end architecture is data that defines the current state of the application. It may include information about the current values ​​and state of components, user actions, selected options, and other entities that affect the behavior and display of the application.
Application state can be stored and managed in a variety of ways, depending on the chosen architecture and the tools used. Here are some of the common approaches:
1. Component local state: Each component in an application maintains its own local state. This state is available only to this component and is used to display and manage its internal state.
2. Global state: One global state is stored in a central store and is available to all components in the application. Examples of such approaches include Flux, Redux, and MobX. Global state allows components to retrieve and update data from a shared store, keeping the data consistent and in sync.
3. Routing state: In some applications, state keeps track of the application's current route. This allows you to determine the current URL path, route parameters, and other routing information.
4. Server State: In some cases, application state may be stored on the server and the client application communicates with the server to retrieve and update the state.
Application state management is an important part of front-end development and can include various patterns and tools. This allows you to create more complex and dynamic applications by ensuring that data and updates are consistent across different parts of the application.