JS: Advanced Testing 1
Jest has a toThrow matcher that allows you to check if a function throws an exception. This matcher can be used to test situations where a function should throw an exception.
The toThrow matcher is very useful for testing exceptions and making sure that a function works as expected when errors occur.