JS: Advanced Testing 6
This test verifies that the getUserMainLanguage function correctly determines the user's main programming language 'MikkeyVespa'. A fake client is created that returns the predefined data. The getUserMainLanguage function is then called with the specified user and client, and the result is checked with expect. In this case, the base language is expected to be 'javascript'.
The second test checks the behavior of the getUserMainLanguage function when the user 'MikkeyVespa' has no knowledge of programming languages. A fake client is created with empty data. The getUserMainLanguage function is then called and the result is expected to be null.