JS: Advanced Testing 7
Monkey patching and the nock library are two different approaches to change the behavior of the code during testing, but with some differences.
Munking patching is a technique that allows you to temporarily change or extend the functionality of existing objects or modules while your code is running. This is achieved by replacing or modifying methods or properties of an object with others to change their behavior for testing purposes. Macking patching is usually used to replace dependencies or emulate some external condition that is difficult or impossible to recreate in a test environment.
nock is a standalone library that provides facilities for simulating HTTP requests and responses during testing. It allows you to create mock servers that intercept and respond to HTTP requests, emulating the behavior of a remote server. nock is typically used to test code that interacts with external APIs or services.
The key difference between mocking patching and nock is that mocking patching is used to change behavior within a code or module, while nock is used to emulate remote requests and responses at the HTTP layer. Both approaches can be useful in testing different aspects of applications, depending on specific requirements and situations.