JS: Polymorphism 1
Parametric polymorphism in JavaScript refers to the ability of functions or types to operate on different data types without being tied to a specific type. It allows you to create generic or generic functions and types that can be used with different data types while maintaining the same structure or logic.
In JavaScript, parametric polymorphism is achieved through the use of generic data types or templates (generics) and functions that can accept and work with arguments of different types. Unlike statically typed languages ​​such as Java or C++, JavaScript implements parametric polymorphism dynamically because JavaScript is a dynamically typed language.

Benefits of using parametric polymorphism in JavaScript:
1. Common logic for different types: Parametric polymorphism allows you to create generic functions or types that can be used with different data types. This avoids code duplication and allows you to use the same logic for different types.
2. Greater flexibility: Parametric polymorphism allows you to write more flexible and reusable code. Functions or types can be written without being tied to a specific data type, making them more abstract and generic.
3. Type safety improvements: Parametric polymorphism helps provide type safety because the compiler or interpreter can check for correct types at runtime. This can help prevent type errors and improve program reliability.