JS: Polymorphism 2
Key (or data) scheduling is a concept that refers to polymorphism and allows you to select different behavior or processing based on the value of a key or data.
In the context of polymorphism, key dispatch is often implemented using conditional statements or switch statements. Depending on the value of the key or data, the program determines which block of code or which function to execute.