JS: Polymorphism 3
Function-key dispatching (or function-type dispatching) is an approach where a specific function is selected and invoked based on the value of a key or type. This allows you to perform different actions or operations based on the key value or function type.
In JavaScript, function key dispatching can be implemented in a variety of ways. One is using an object, where the keys are the values ​​to dispatch on, and the values ​​are the corresponding functions.