JS: Polymorphism 4
Dispatching by file name (or by file extension) is an approach where different actions are performed depending on the file name or extension. This allows you to associate certain operations or functions with certain types of files in order to process them in a specific way.
In the JavaScript programming language, filename dispatching can be implemented using conditional statements (such as if-else or switch-case) or using an object where the keys are filenames or extensions and the values ​​are the corresponding functions or actions.