JS: Polymorphism 5
Duck Typing is the concept of polymorphism in programming languages ​​such as JavaScript. It determines the type of an object not from its explicit type or class, but from its methods and properties. If an object looks like a duck, swims like a duck, and quacks like a duck, then it can be considered a duck, regardless of its actual class or type.
What matters in duck typing is not inheritance or interfaces, but the sharing of common methods and object properties to achieve the desired behavior. If an object has methods and properties that match the expected interface, then it can be used in place of another object, regardless of its actual type.
An example of duck typing is the use of different objects in JavaScript that can be used to do the same thing as long as they have compatible methods and properties. For example, if we have duck and robotDuck objects, and both objects have a quack() method, then we can easily call quack() on each of them, even if they have different types or classes .