JS: Polymorphism 6
Null Object Pattern in JavaScript is a design pattern that suggests using a special null object as a placeholder instead of checking for null or undefined values. This null object should have the same interface as other objects in the system but its methods should have empty or neutral implementations.
When using the Null Object Pattern, instead of checking for null or undefined values for objects, we can simply pass the null object. This simplifies the code and eliminates the need for explicit null checks before calling object methods.
The advantages of the Null Object Pattern include:
Code simplification: There is no need for explicit null or undefined checks, which simplifies and shortens the code.
Prevention of runtime errors: Using the null object ensures that methods are always called and prevents exceptions such as "Cannot read property of null".
Support for the Open/Closed principle: If we want to add new implementations of objects, we can simply add new classes that implement the null object interface without modifying existing code.
In JavaScript, the Null Object Pattern can be implemented by creating a null object class with empty or neutral method implementations. This class should have the same interface as other objects in the system so that it can be used as a replacement for real objects.