JS: Polymorphism 7
Create a User class with a constructor that takes a name parameter. Inside the constructor, set the name property to the passed argument.
Add a getName() method to the User class that returns the value of the name property.
Add an isUser() method to the User class that always returns true, as objects of the User class are users.
Create a Guest class without any parameters. In the constructor, set the name property to 'Guest'.
Add a getName() method to the Guest class that returns the value of the name property.
Add an isUser() method to the Guest class that always returns false, as objects of the Guest class are guests.
Export the getGreeting function as the default export in the helpers.js file.
In the getGreeting function, accept a user argument.
Inside the getGreeting function, call the isUser() method of the user object. If the method returns true, return the string 'Hello <Name>!', where <Name> is the result of calling the getName() method of the user object.
If the isUser() method returns false, return the string 'Nice to meet you Guest!'.