JS: Polymorphism 8
Starting with the WeatherService class. In this class, we create a constructor that takes an HTTP client (httpClient) as a parameter and saves it inside the class instance.
In the lookUp method of the WeatherService class, we accept the city name (cityName) as an argument. Then, we construct a URL using the base URL for the weather API (apiUrl) and the city name.
Next, we send a GET request using the HTTP client (httpClient) to the formed URL and receive a response (response) from the API.
After that, we parse the JSON response (response.data) to get an object with weather data for the specified city.
The lookUp method returns an object with weather data, containing the city name (name) and temperature (temperature).
In the bin/weather.js file, which is your console application, you:
a. Import axios to create an instance of the HTTP client, which will be passed to the WeatherService.
b. Create an instance of WeatherService, passing the HTTP client (axios) as an argument to the constructor.
c. Extract the city name from the command-line arguments (in this case, process.argv[2]).
d. Call the lookUp method on the WeatherService instance to get the weather data for the specified city.
e. Print a message on the console with the temperature in the format "Temperature in <city name>: <temperature>°C".
Now, when you run the application with the command node ./bin/weather.js berlin, for example, it will send a request to the weather API, get the weather data for Berlin, and display a message like "Temperature in berlin: 11°C" (for example, if the temperature is 11 degrees Celsius).