JS: Polymorphism 9
Strategy is a design pattern that allows you to choose an algorithm or strategy for performing a certain action on the fly, without the need to change the main class. This is done by separating the algorithm into a separate class and injecting it into the main class, which makes it easy to replace the algorithms without affecting the main class code.