JS: Polymorphism 12
The generatePassword function takes two arguments: length (the length of the password) and options (an array of options for generating the password). It creates the defaultOptions object containing default options (symbols, numbers, uppercase) with values set to false.
Next, the function uses the Object.fromEntries method to transform the options array into the preparedOptions object, where each element of the array becomes a property of the object with a value of true. This allows easy tracking of which options were passed to the function and sets their corresponding values to true.

Then the function merges the defaultOptions and preparedOptions objects using the spread operator (...) into the finalOptions object. This ensures that all options are taken into account, and any options that were not passed will have a value of false.
Finally, the function calls the generator.generate method and passes it the finalOptions object to generate a password with the specified options.