JS: Polymorphism 13
The State pattern in JavaScript is a behavioral design pattern that allows objects to change their behavior based on their internal state. It helps to avoid large conditional statements associated with different object states and separates the state logic into individual state classes.
The main idea of the State pattern is to create classes representing different states of an object and an interface (state interface) for interacting with these states. The object then holds a reference to the current state, and all operations are invoked through the state interface. When the object changes its state, it switches to a different state class.