JS: Frontend Architecture 7
We create a state object (Model) named "state" that will contain information about the current state of tabs in the form of an object "lists".
We create the "onChange" function from an external library, which tracks changes in the "state" and performs specific actions, passing the current and previous tab states.
We create the "watchedState" object, which acts as a proxy for the "state" object and allows us to track its changes.
We get all tab elements with the attribute "role='tablist'" using "querySelectorAll".
For each "list" element with the attribute "role='tablist'", we add a click event listener.
When a tab is clicked, we update the "tabId" value in the "state.lists[listId]" object using the "watchedState" proxy. This update triggers the "onChange" function and changes the active tabs accordingly.
In the "onChange" function, we retrieve the current and previous tabs based on their identifiers and add/remove classes to set the active and show states for the corresponding elements, thus toggling the tabs when the state changes.