JS: Frontend Architecture 8
We create a state object (Model) named "initialState", which will contain information about the registration process state (signupProcess) and the form state (form). We also define an "elements" object with references to DOM elements.
We create the "handleProcessState" function, which is responsible for updating the View based on the registration process state. In this function, we use DOM elements from the "elements" object to change the container content and the submit button state depending on the process state.
We create the "handleProcessError" function, which will be used when errors occur in the registration process, but this function is not implemented in the provided code.
We create the "renderError" function, which is responsible for displaying error messages near the corresponding form fields.
We create the "renderErrors" function, which updates the View when validation errors occur in the form. This function iterates over the form fields and updates classes and text messages for the corresponding fields based on the presence or absence of errors.
We create the "render" function, which will be used to track changes in the state (Model) and update the View accordingly. In this function, we use a "switch" statement to determine the changed path in the state and call the appropriate handling functions for the changes.
We export the function that will be used as a controller (Controller) to manage state (Model) and view (View) updates based on user actions. In this function, we set event listeners for the form fields and the submit button, which will update the form and signupProcess state and call the "render" function to update the View.