JS: Frontend Architecture 9
We create an array "languages" with the available languages for the language switcher.
Define the "handleSwitchLanguage" function, which takes the application state "state" and returns an event handler function. This function will be used to change the current language when the language switch button is clicked.
Create the "render" function, which takes a container, the watched state object "watchedState," and the "i18nInstance" object for localization. In this function, a block with buttons for language switching and a counter for click tracking is created. Event handlers are also added to the language switch buttons and the counter button.
Export an asynchronous default function, which will be called when the application is launched. In this function, an instance of the "i18nInstance" object is created for localization. Then, this object is initialized and started, passing the default language and localization resources to it. An object "state" is also created with information about the current language and the number of clicks.
Create a handler for the watched state changes using the "onChange" function. Inside this handler, we use a "switch" statement to handle changes in the state. When the language is changed, we call the "changeLanguage" function of the "i18nInstance" object and re-render the view using the "render" function. When the click counter changes, we also re-render the view.
Finally, we call the "render" function to initialize the view.