JS: Diving into classes 2
Import Statement:
The code starts with an import statement that imports the AppError class from the file './errors/AppError.js'.
Function anyErrorCatcher:
The code defines a function anyErrorCatcher that takes two parameters: errorHandler (a function to handle errors) and errorInstance (an error class to check against).
Inside the function, it returns another function that serves as an error handler.
The inner function takes an error as its parameter (representing the error that occurred).
It checks if the error is an instance of the provided errorInstance class.
If it is, it calls the errorHandler function with the error as its argument.
If not, it throws the error to be caught elsewhere.
Function appErrorCatcher:
Similar to the anyErrorCatcher function, the appErrorCatcher function takes errorHandler and errorInstance as parameters and returns an inner function (error handler).
The inner function also takes an error as its parameter.
It checks if the error is an instance of both the provided errorInstance class and the AppError class.
If both conditions are met, it calls the errorHandler function with the error as its argument.
If not, it throws the error.
Function customErrorCatcher:
Similar to the previous functions, customErrorCatcher also takes errorHandler and errorInstance as parameters and returns an inner function (error handler).
The inner function takes an error as its parameter.
It checks if the errorInstance is falsy (null, undefined, false, 0, empty string) and whether the error has a property isCustomError set to true.
If these conditions are met, it calls the errorHandler function with the error as its argument.
If not, it throws the error.
Additional Import Statement:
The code again includes the same import statement for the AppError class from the file './errors/AppError.js'. Note that this is a duplication in the given code.
Usage of Functions:
After defining the functions, they can be used to create error handlers that handle specific types of errors.
Each function can be passed an appropriate errorHandler function and an errorInstance class to check against.
When an error occurs, the corresponding error handler will be executed if the conditions specified in each function are met.